Added CMSHome to routing and removed test code
In this update, the CMSHome component was added to the application's routing. This allows users to navigate to the CMS home page through the app's navigation system. Additionally, some testing code that was previously used for checking the functionality of CMSHome has been cleaned up and removed from App.jsx.