Refactor 'OurCuts' to use local storage
The 'OurCuts' component has been refactored to utilize local storage. Previously, the cuts data was hardcoded into the component. Now, it attempts to fetch this data from local storage upon rendering. If successful, it sets the state with this data. The display logic remains unchanged. Additionally, a conditional rendering for cut images has been added: if an image URL is available in the cut data, it will be displayed; otherwise, a default scissors icon will be shown.